Set event on fatal exception in log worker thread
When a flow run finishes, it flushes the log handler and blocks
until the workers report that they have finished sending logs.
If the log worker failed outside of the portion that makes the
request to the API, the exception would kill the thread without
setting this event. Consequently, the flow run would hang and
run forever.

This pull request logs the exception (matching our other log
worker exception handler) and uses a `finally` block to ensure
that the event is set on failure.

Issue raised in  https://github.com/PrefectHQ/prefect/issues/5699

This pull request does not address the underlying event loop error
raised in that issue. We will need to avoid spawning event loops
in a thread to resolve that.