Bump full test suite timeout to 30 minutes
We appear to be bumping into the upper limits recently. If something is hung and running forever, I want it to be very clear that was the case. In these recent cases, it is not clear if the test suite was hung or just did not finish fast enough. We can lower this again following #1885 if necessary.