Wrapping up the CLI for creating new Deployments and legacy DeploymentSpecs.
This allows for a future use case where we'll let users package flows ahead of
time and use the packaged flows for deployments.  If we recieve a
`PackageManifest` as the `flow=`, we will simply use it without attempting to
deserialize the flow from it.

Note: this is still a WIP, the `prefect deployment preview` command and tests
are still failing.