Bypass the task runner when submit is not used
This fulfills the requested change from @cicdw to reduce serialization/remote execution errors while not opting into task runner usage. A future change from task runners -> workers should improve the type passed through. Right now, we create a new sequential task runner on each call and allow it to be disposed afterwards — this is a bit of a short-term hack.