Initial draft of a more forgiving inner loop for long-running services.
This takes the approach of remembering the recent track record for a service,
so that momentary blips won't cause it to exit prematurely, but an ongoing
series of problems that might indicate a deeper issue (network, server, or cloud
outage) will.  This is nice because it will exit immediately if the first call
fails, and then once it "ramps up" and has a proven track record, it will be
able to overcome short blips.

Initial sketch, without a test suite yet, just to get some feedback on the idea