Load flows for deployments in a worker thread
If not run in a worker thread, it is executed in an "async" context which means that any methods decorated by `sync_compatible` will return a coroutine. However, top-level code in a script is always intended to be run in a "sync" context, so the user can never `await` the coroutines and code that works without a deployment will break while using a deployment.