Automatically publish documentation on release (#6459)
Instead of just opening a pull request, the release automation will merge the pull request. This will not occur if there are merge conficts. Previously, if a maintainer squash merged the pull request the history of the `docs-deploy` branch would diverge from `main` resulting in failure of this workflow on the next release.