Improve performance of flow run serialization.
In testing we noticed performance issues when using
jsonable_encoder with large, nested data structures.
FlowRun.parameters is often such a structure.

Not only can it be slow, but FastAPI does not run the
call in a worker thread, so it blocks the main event
loop.

We also now use `orjson` by default for our pydantic
model serialization.

In testing, this approach resulted in a 75% reduction
in CPU time needed to process deliberately problematic runs.