Prevent non-UUID slugs from raising errors on the BlockDocuments APIs. (#6541)
In Prefect Cloud, we observed some errors when clients would send requests for
`.../block_documents/null`, which should really be handled at the routing layer
with 404s when the path UUIDs can't be parsed.

Note that this is just correcting the server-side issue, but does not attempt
to diagnose the client-side issue at this time.  Also, this does not attempt
to go through every route in Orion that includes UUIDs in its path.