Update agent to prevent crash when infrastructure errors after submission (#6456)
* Fix error message in `submit_run`

Refers to out of date "flow runner" concept

* Update tests

* Update agent to prevent crash when infrastructure errors after submission

Running infrastructure is a two step process:
- Submission: infrastructure is created for the flow run
- Monitoring: the infrastructure is watched for success/failure

AnyIO's `TaskStatus` is used to report the completion of submission
allowing a caller to use `task_group.start` to await submission without
blocking on monitoring. If an error occurs during submission, the
agent captures it and marks the flow run as failed. However, if an
error occured during monitoring, the agent does not capture this
exception. Instead, it is raised in the agent's task group which
can cause the agent and any other concurrent tasks to fail.

A single flow run should not be able to crash the agent. To resolve
this, the agent is updated to capture exceptions during the
monitoring phase as well. In contrast to submission, we do not mark
the flow run as failed as this could be unrelated to execution of
the flow itself.

* Update implementation to be robust to task status failures