Proactively look for duplicate blocks instead of relying on `IntegrityError`.
On PostgreSQL, relying on the `IntegrityError` to indicate a constraint
violation does work, but leads to a lot of noise in the PostgreSQL logs.  Also,
tracing libraries see these spans as errors, even though they are normal ops.

This is a backported fix from Prefect Cloud and fixes PrefectHQ/nebula#2165