Restores proxy-awareness for the Orion client for remote Orion servers.
We were overriding the `transport` on our Orion connections in order to set the
number of retries.  When doing this, we would need to include all of the logic
for how to discover if the user has proxies configured in their environment.
Instead, just set the retries _after_ the fact.

Fixes PrefectHQ/nebula#2356