Add dark mode to docs
This commit will add dark mode to the docs.

Material for MkDocs has a built-in dark mode called slate, and supports
automatic light/dark mode selection based on user preference.

This commit will add automatic light/dark mode selection, as well as a
toggle icon in the top nav bar for switching between the two.

Color schemes are evaluated in order top to bottom. Light mode is listed
first, so the default will remain light mode, as it is now.