Introduces Prefect Events and the clients that can emit them
This introduces the shell for what will become a more fleshed-out implementation
later.  In this change, I'm just carving out the `prefect.events` namespace,
introducing the client-side `Resource` schemas and abstract client for emitting
Events.

This includes a FastAPI dependency that we can use in API routes to have access
to an event client.  The idea is that in the beginning, this will be a
`NullEventsClient`, so there should be only a very tiny performance change to
use it.  Later, we can replace the implementation using FastAPI's
`dependency_overrides` to drop in a "real" client.