Allow `reject_transition` to use current state
Previously a REJECT response required a new state which was then always
written to the database. Here, the `OrchestrationRule.reject_transition`
utility is updated to accept `state=None`. When this is used, the
REJECT orchestration response will contain the current state of the run
and a new state will not be appended to the run's states. This allows
REJECT semantics to be used to prevent transition to a proposed state
without providing an explicit alternative. This is useful when a client
is proposing a state with the assumption that they are in a different
starting state than the orchestrator has.