Annotations in KubernetesJob
Adds the ability to add an optional annotations dictionary to the
KubenetesJob class. This allows users to specify annotations to be added
to their flows. An empty annotations dictionary will be injected if none
is specified by the user, but it does not have to be included in the
minimal manifest.