Add test for idempotency key flow run filter
This commit adds a pytest fixture for generating a list of FlowRuns that
have two different idempotency keys, so that it can be shown that the
filter retrieves one and not the other.

In addition, this commit adds a test that we can filter flow runs based
on their idempotency key value.