feat: generate flow_run_name within flow_run context (#8895)
We accept function for flow_run_name but it currently cannot access the
flow run context. So the function will not be able to use the parameters
passed to the flow, the flow name or other useful contextual
information.

Moving it under the flow run context will enable it to utilize this.

There will be regression if anyone was depending to directly access
parent flow run context

Issues:
  * https://github.com/PrefectHQ/prefect/issues/8895