fix type checking of Flow.__call__ and Task.__call__
The first two overloads for `Flow` lead to incorrect type inference with mypy and pyright, while simply leaving them out works.
The same is true for `Task`, although this only has one faulty overload.

In particular, consider this overload:
```python
    @overload
    def __call__(
        self: "Flow[P, Coroutine[Any, Any, T]]", *args: P.args, **kwargs: P.kwargs
    ) -> Awaitable[T]:
        ...
```
This confues pyright which get's tripped up by the `Coroutine`/`Awaitable` change. Instead, the plain
```python
    @overload
    def __call__(
        self: "Flow[P, T]",
        *args: P.args,
        **kwargs: P.kwargs,
    ) -> T:
        ...
```
is equally valid for sync and async functions. The in the latter, the T is an `Awaitable[U]`, for some type U.

The other overload that this commit removes is also incorrect in general:
```python
    @overload
    def __call__(self: "Flow[P, NoReturn]", *args: P.args, **kwargs: P.kwargs) -> None:
        ...
```

Consider this:
```python
@flow
def foo() -> NoReturn:
    while True:
        pass

foo()
print("This is unreachable, but the overload implies that foo() eventually returned None")
```
Again the plain overload would preserve the correct types.