chore: Improved code readability in casting environment variables for known mocked deployment runtime attributes.
- Using a type casting dictionary
- Include unit tests for each allowed runtime attribute

Furthermore, if key type is not in casting dict, it raises an error. It is a mechanism to avoid mocking attributes
for complex types (lists, dicts, ...), which would increase the complexity of the code and would be more difficult to use by the user.

Finally, an unused function has been removed from src/prefect/runtime/flow_run.py file.