Add hint to "Invalid timezone" error message
If the error message does not appear to be a valid IANA tzdata
timezone specifier, then we return an "Invalid timezone" error,
but do not provide any hints to resolve.

This change updates the error message and documentation string
to indicate that the timezone is to be specified in IANA tzdata
format, such as America/New_York.

Closes: #10006