Adding explicit tests for the format of the `PREFECT__FLOW_RUN_ID` env var
We had an incident where we caused a large number of 404s because we changed the
read_flow_run route to use the specifier {id:uuid}. This specifier expects
UUIDs to have dashes, but the PREFECT__FLOW_RUN_ID variable is produced with
flow_run.id.hex and will be a 32-character hex string without dashes.

This test makes it explicit that we're supporting that unexpected format so we
don't accidentally break this again.