Vendoring FastAPI as a first step toward `pydantic>2` support
While we begin to work on `pydantic>2` support, we need to support living in
any of three different "worlds" concurrently:

1. `pydantic<2` is installed (today)
2. `pydantic>2` is installed, `prefect.server.schemas` are pydantic v1 models (intermediate)
3. `pydantic>2` is installed, `prefect.server.schemas` are pydantic v2 models (end state)

FastAPI's compatibility layer only expects (1) or (3), based on the reasonable
assumption that FastAPI users are in control of their `pydantic` version and
schemas.  However, `prefect` will need to exist in world (2) for a prolonged
period of time while we deprecate support for `pydantic<2`.

To this end, we are removing our dependency on the public `fastapi` package, and
vendoring it privately to `prefect` for use exclusively for `prefect.server` and
the short-term login server used for Cloud login.  This means that end users
may have any version of FastAPI they like without conflicting with ours.  We
will make some adjustments to the vendored version's compatibility layer to help
bridge the gap between worlds (1) and (3).

Long-term, we will spin this down and replace it with the public `fastapi`
package again.