Isolating the server-side schemas from the client-side schemas
As the next step toward `pydantic>2` compatibility, we're separating the
implementation of the `prefect.server.schemas` from the `prefect.client.schemas`
by duplicating some of the shared base classes/fields/utilities.  This will
allow us to evolve the client schemas to support `pydantic>2` at a different
rate than the server schemas.

This is intended to be a purely no-op refactoring, and the next step will be
to strip down some of the implementation complexity in the server schemas in
preparation of `pydantic>2`.