Introducing a tool to help with bulk tasks over Prefect-maintained collections
There are occasions where we need to make bulk updates across all of our
collections (currently ~30 repos now), such as dependency bumps, adapting
to an upstream change, or to re-apply changes from the collections template.

In these cases, it can help to have a tool that operates over each repo in a
uniform way.  This script is a starting point with shortcuts for those kinds of
tasks, and should be pretty easy to extend to other use cases.

Note that this script is currently opinionated about a couple things that may
not be true for everyone on the team:

* That you follow the `GOPATH` style of code repos on your local machine:
  `~/src/<site>/<owner>/<repo>``, like `~/src/github.com/PrefectHQ/prefect`
* That you use `pyenv` to manage multiple python versions and virtual
  environments for local development