Always use the larger `oss-test-runner`
We're discovering that our test suite requires the larger test runner to
complete in a reasonable time, and that we are starting to see widespread
failures when running on the default runner.  This is due to an unclosed
connection, which is likely due to internal timeouts in the test suite.

This removes the internal user constraint entirely, because it appears that the
suite won't complete successfully on the default runner anymore.  This will give
a better experience for external contributors.  If we see signs of abuse, we can
require approvals for new contributors.

Note: I'm also adjusting the `pydantic` requirement due to how the
`generate-lower-bounds.py` script works.  We had a redundant `>=1.10.0,>=1.7.4`
line (from recent compatibility work) so the lower bound was generating as
`1.7.4` (which we don't support).