Exclude any flow/task run with a `NULL` `start_time` from Graph v2
There can be a brief period of time when a flow/task run has been created but does not
yet have a `start_time` or even an `expected_start_time`.  This is causing a pydantic
`ValidationError` when trying to read the graph.  We can safely ignore these
cases, as they can't be plotted on the graph visualization anyway, and the runs they
represent will "pop in" as soon as they have a valid time.