Forgetting the `args/kwargs` to a Prefect `Call` after they are used
We use an internal object called a `Call` to capture the function,
`args/kwargs`, and some contextual information about function calls that are
potentially crossing concurrency boundaries.  These objects are used to track
calls and make their results available in a sync/async agnostic way.

When a user task is run, it is ultimately done so with a `Call` created by
`begin_task_run` that captures the task arguments.  That `Call` seems to still
be referenced by the event loop between flow runs, and if large objects were
passed as arguments, the garbage collector can't free them.  This change breaks
that reference cycle so that after the function of a `Call` has been invoked in
the right concurrency context, the `args` and `kwargs` are set to `None`, as
they can't be used meaningfully anymore (except for debugging).

Fixes #10952