A avoid rerunning task runs forced to Completed state
We want to avoid rerunning task runs in a Completed state if they’ve been
"forced" into that state from a Failed or Crashed state. This is because it may
be unlikely or impossible for the failed task run to succeed again on a
subsequent retry, and in that case, a user may want to force the run into
Completed state and allow a flow run retry to run the other tasks in the flow.

To achieve this, we assign a placeholder result during forced state transitions
from Failed or Crashed to Completed. Then, if we receive another proposed
Running state for that task run, we'll reject the transition with a Completed
state, the same as we would if the task had actually completed successfully and
persisted a result. This causes the client to treat the task run as Completed.

fixes #10749