Use default values if no run input was provided
It's possible to have a flow run waiting for input that uses only
optional fields wtih defaults, especially if the input class was
created with `RunInput.with_initial_data`. In this case, we should
pass the empty data to the `RunInput` class and rely on validation
to ensure any required fields are present instead of rejecting the
empty input.