Implementing task subscription filtering
The goal here is for each subscriber to be able to filter to a subset of the
tasks they are interested in (i.e. which tasks a task server is serving).  This
required routing task runs into queues and monitoring multiple queues for
tasks.  This was a good opportunity to refactor some of the mechanics here to
capture the idea that a `TaskQueue` is really two queues (a "regular" queue and
a high-priority retry queue).  This led me down a path of tidying up the
implementation to remove globals and encapsulate more of the behavior.