Making the flow timeout tests deterministic
Our prior versions of these tests depended on how fast the test runner was,
which led to a significant number of timing flakes here.  The changes here make
these tests deterministic (with `pytest-timeout` as a backstop).

Note: the tests have also changed to acknowledge that we can't currently
interrupt a long-running sleep within a flow.