Reducing the dimensions of the unit test matrix
In order to get cycle times for PRs down to reasonable levels, we're temporarily
dropping the PostgreSQL 13 and pydantic v1 tests.  We may be able to bring these
back in the future, but for now we need to focus on stability and cycle times.

I am also removing an apparently dead branch around installing the lower bounds
of all requirements.  We have support for this, but we haven't had it in the
matrix for some time.  Removing it for clarity.