Disposing of all database engines created during a test session
One of the persistent flakes we've seen are lingering connections to the
postgres database, which may manifest as a `ResourceWarning` for a socket to
5432, or as a SQLAlchemy warning because an async connection can't be closed
with `__del__` (i.e. garbage collection).

I detected that under high concurrency and load, we can sometimes leave a single
engine open and not disposed.  In my testing it was always an engine that had
been created by `apply_migrations`, because it was running from a separate event
loop on a separate thread.  Attempting to correct the problem by disposing the
engine in `apply_migrations` lead to other more cryptic errors, so I'm going to
do this as session teardown here.

This should also resolve the issue that pluggy was warning us about.

Fixes #11820