Tracking connections process-wide for test assertions
This adds a tracking utility that observes all connections produced by our
engine's pools, allowing us to make assertions and clean them up at the end of
a test session.  This utility helps to avoid a `ResourceWarning` about
connections being cleaned up with `__del__` instead of being deliberately
closed, because it keeps a reference to all connections (open or closed) until
we can dispose of the engines and all the pooled connections deliberately.

The utility is only active during unit tests, when `TRACKER.active` is set, so
we shouldn't have concerns about memory use in long-running servers.

Fixes #11820