Bring back the task subscription API tests
These had been disabled while we were looking for test flake culprits, but they
should be okay to run.  The original implementation running an in-process server
was far too fragile.  This implementation uses the `TestClient` and is an
all-sync test suite using starlette's socket testing utilities.

In retrospect, many of the hangs I was experiencing the first time were likely
due to protocol errors, where both the client and server were sending/receiving
and not lined up with each other.  It took a bit of care to get those all lined
up.