Run the full test suite in each leg of the matrix
We had been splitting off a set of tests we called the "services" tests, which
includes running some more invovled tests with the process and docker infra.
This only splits off 53 tests from the 8k test suite, so the overhead of a full
separate leg of the checks seems unnecessary.  We aren't skipping them for the
pydantic, sqlalchemy, and datadog tests, and they all seem to run just fine as
a unifed suite.