Correct a memory leak with the outbound task run websockets
I did some independent testing of this with FastAPI and an `asyncio.Queue` in a
simplified environment.  I found that when waiting for the next message on an
empty `asyncio.Queue`, the server will not notice that the client has
disconnected.

With a brief timeout on the `queue.get()`, the disconnects will be noticed and
the server can clean up that socket handler and anything related to it properly.