Carving out room in the test suite for server-side event tests.
As we start introducing more server-side event features, we'll want a place for
those tests.  However, we'll also want a place to be able to test _all_
events-related features in one shot, so I'm creating the structure:

tests/events/client/
tests/events/server/

I know this is a departure from `tests/<feature>` and `tests/server/<feature>`,
but being able to test all of the work together quickly is worth the benefit.
If we end up liking this, we may want to bundle up the other features that have
both a client- and server-side footprint this way as well.