Refactor the DeploymentTrigger into its own hierarchy to match Trigger
As I was starting to write more tests for DeploymentTrigger, I realized that
I was effectively reproducing tests for validation that Pydantic would do for
us if we just matched the hierarchy correctly.  Unfortunately there's no
way to inject a few fields into a base class for another class hierarchy, so I'm
reproducing the full hierarchy and its validate rules here.