Adding more tests for events schemas to increase coverage
In Prefect Cloud, I took the approach of letting these utilities be tested by
the calling code that needs them (like in triggers, storage, queries, etc), but
because these will form a first-class user-facing API, I wanted to test all of
the public interface directly.  This brings the coverage of the events modules
up to 95%.  I was having some trouble with `#pragma: no cover` on some
conditional imports, so they are still reporting as uncovered for now.