Introducing an internal orchestration API client for use in events
In loosely-coupled areas of the system, it can keep things more contained if we
use an HTTP API call rather than direct model calls.  The events and automations
subsystem will use this for executing actions and rendering templates, and the
presence of this internal client will make porting from Prefect Cloud easier.

Note: I'm bringing a few arbitrary methods over from Prefect Cloud, but the idea
is that we'll bring more over as needed.  These are just to give us something
to anchor the test suite to.