Introduces the server-side automations schema models
This introduces just the models, including the instance methods, but doesn't
add much in the way of new tests because most of this code is tested as part of
the triggers services.  I need the schema models available so I can build out
the ORM and CRUD APIs.

I debated whether to cut these instance methods for better coverage or to keep
them in and accept the coverage drop.  Keeping them in means less churn when we
do add the triggers system, so I'm good with that.