Adding the ORM classes and model modules for Automations and it's state tables
This ports the Prefect Cloud implementation and test suite of the storage for
automations.  This also includes a sprawling change to eliminate all the uses
of the root-model constructor `SomeClass(__root__={...})` with an intermediate
form `SomeClass.parse_obj({...})`.  When we're ready to go all-in on Pydantic 2,
we can change those to the much cleaner `SomeClass({...})`.