Adding an Automations CLI
This CLI supports listing, inspecting, pausing, resuming, and deleting
automations against Cloud, and against the OSS Prefect server when the
experimental flag is on.

Note I'm also consolidating some tests and files to keep as clean of a coverage
target for events as possible while we're still heavily developing on it.