Renaming `prefect-cloud.*` events and labels to `prefect.*`
While introducing automations to open-source, we want to be consistent with the
naming of events, since all orchestration domain concepts use events and labels
like `prefect.*`.  While automations started life in Prefect Cloud, they are
now going to be considered orchestration domain objects.  This will break a
conceptual compatibility for the near term: if you're triggering on
`prefect.automation.*` events in OSS, you'll need to change that to
`prefect-cloud.automation.*` in order to deploy that to Prefect Cloud.  We're
good with that tradeoff until we can get the concepts in Prefect Cloud renamed
as well.