Add support for using the "normal" Trigger classes for flow.serve and .deploy
The `Deployment*Trigger` classes help capture some additional information about
the automation to create (name, description, whether it's enabled, etc) or
about how to customize the `RunDeployment` action, which can be useful for
stored triggers in a `prefect.yaml`.  When serving flows dynamically or when we
don't need as much customization, we can also accept triggers from the "normal"
`Trigger` hierarchy.

This also makes the `Deployment*Trigger` classes subclasses of their respective
`Trigger` types to avoid duplicating fields and validations.