Preparing to introduce `mypy` checking, starting with client-side events
We have a long-term goal to have the entire Prefect library pass mypy in strict
mode, and the journey of a thousand miles starts with a single step.  This
resets the `mypy` configuration we had and starts with a couple of modules we've
identified to tackle today.  This includes the client-side events code.

Probably the most controversial thing I'm doing here is making `DateTimeTZ` just
an alias to `DateTime` and then dynamically adding the appropriate `pydantic`
validator functions to it.  This allows us to avoid a giant swath of typing
errors where it looks like we're downcasting a `DateTime` to a `DateTimeTZ`.

This change has been in effect on Prefect Cloud for some time now, so I'm pretty
confident that it is a drop-in replacement for what we were doing before.