`mypy` coverage for the rest of client-side events
The original mypy file selector was excluding files in deeper directories.  This
simplifies it to just include the directories we're able to check types on.

I also took a pass through all the low-hanging fruit in `tests/events/client`,
but we'll need to do some more work on the fundamentals first.