Migrating all uses of the banned characters validation to a self-validator
There were a few sub-flavors of this, but I was most surprised that we allow
for `Deployments` to have empty string names.  I'll track that separate issue.
My aim here was for no behavioral changes except for the new normalized
validation message from Pydantic (and the 422s instead of 400s).