Fixing stray files from the test suite, with a utility to notify us about them
We've had trouble in the past with tests inadvertently writing files to the
current working directory instead of isolated to a temporary directory.  It's
far too easy to introduce these, and they can be frustrating to track down.

This adds a module-level autouse fixture to alert us when a test suite had
dropped new files in the working tree.

Counterpart of #13380