Replace `Block`'s `_iter` implementation with a model serializer
`Block` needs to emit a `block_type_slug` field any time it is serialized in
order to enable type dispatch (looking up subclasses of the base Block class),
but it was doing this with `_iter`, which is no longer in play.  Using a
`@model_serializer(mode="wrap")` allows for easily adding things to the
serialized output.