Align the first and second scripts
The first iteration of the script varied a bit from the
second iteration, but the highlighted lines in the second
iteration suggested only two lines had changed. This isn't quite
true - for example, the function signature of 'get_repo_info' changed
to include two new arguments.

When following along with the quickstart, this makes it look like a
"patch", but after applying the same changes to my code, I obviously
got a different result.

This change aligns the first version with the second. A drawback is this
makes this initial script longer, but:

- It's still not a very long script anyway.
- The "copy" button is still there, which I suspect most people use
anyway.
- Those following along can now actually just change 4 lines now and get
the new result.