Bringing back the integration tests (#13645)
Here we're simplifying the integration tests and focusing their purpose. In
3.0, these are not cross-version compatibility tests. Instead, they are
tests that we expect to be true against a live running Prefect server of the
same version, or against Prefect Cloud. The purpose of this integration test
suite is pretty much to make sure that the integration flows work against a
live server so we can depend on them for Prefect Cloud compatibility tests.

Prefect Cloud's client compatibility tests are now designed to clone the version
of the tests at the tag that represents the released version, so we no longer
need any version checking in the integration tests.

If we decide later on to re-introduce cross-version compatibility tests, we
won't expect that 3.0 clients will work with 2.x servers, so we can scrub this
test suite now and reboot it when we have future versions of 3.x.