Addressing a race condition with the automations integration tests
One way that these automation tests could fail is if connecting to the websocket
to listen for the triggered event takes longer than it takes to emit the
events and trigger the automation.

By using an `asyncio.Event` to signal when the socket is connected, we can wait
for that signal before we proceed with the rest of the test.