Automating the Mintlify OpenAPI docs
This script will reconcile the current state of the REST API with what's in
the `docs/<version>/api-ref/server` directory, as well as the navigation within
the `mint.json` file that defines the overall documentation.

I'm also wiring this script up as a pre-commit hook so that any changes to the
API and schemas are automatically reflected in the docs.  To make this the least
invasive as I can, this attempts to install the required `@mintlify/scraping`
package if it is not currently installed.

Fixes #13907